requirements.txt
aiohttp>=2.2.5
<<<<<<< HEAD
ruamel.yaml
markdown2
more_itertools
=======
>>>>>>> parent of b8806b4 (config.yaml)
alphabet_detector
beautifulsoup4
bleach
cachetools
CurrencyConverter
emoji
feedparser
future
geopy
git+https://github.com/starry69/python-markdown2.git
#googletrans
gpytranslate
humanize
jikanpy
lxml
nltk
Pillow>=4.2.0
pretty_errors
psutil
psycopg2-binary
pynewtonmath
pyrate-limiter
python-telegram-bot==13.11
regex
requests
speedtest-cli
spongemock
sqlalchemy==1.3.20
pyrogram
tgcrypto
ujson
wikipedia
zalgo-text
cloudscraper
